Refactor existing settings to make things more reusable and match naming conventions
